Fix NotAuthorized issues for cleaning up temp builds
This happens because we are using an authenticated build client to query
through temporary builds that need deletion and if it takes more than 5
minutes, it fails to get a new auth token and fails with NotAuthorized
exception.

This fix makes sure that iterating through the endpoint is done
anonymously to not have to deal with token renewal.